Spread of context into all procedures and removal from structures
In an attempt to make lcms thread safe this is a start in pushing
the context through almost all the functions and removing any
reference to the context from the structures.  In the lcms code
base, the libraries both build as well as a couple of the test
applications.   A bit more work is required to make testbed, tiffdiff,
tifficc, transicc and jpegicc to build.