Replace cmsChangeBuffersFormat with cmsCloneTransformChangingFormats
This removes a thread unsafe function, and replaces it with a thread
safe one that does the equivalent job.

We do this by splitting transforms in two. Each cmsTransform now has
a 'core' part and a part that can be 'cloned' to change the buffer
formats without affecting other clones.

The internal tables are shared, so the cloning operation is very
low cost.