Regularize API, and various tidy ups.
We insist on passing the ContextID to profile/link etc
destructors, and yet provide 2 different constructor
forms. This seems silly. Let's just bin the 'THR' ones
and insist on ContextIDs everywhere.

The rule we follow is that all external API functions
require a ContextID. All internal ones take one if
required. We strip out some cases where we are passing
them around internally for no good reason.

Fix the utilites and testbed apps to actually compile now
(or at least, to fall over due to lack of libs, not due
to errors in the code).

Update the test app to use cmsCloneTransformChangingFormats
rather than the removed cmsChangeBufferFormats.

Fix various warnings (largely due to ContextIDs being passed
and not used).