Avoid problems with OptimizeMatrixShaper mismatching whitepoints.
When using a Lut based optimisation, Lcms specifically checks that
the whitepoint of the input maps to the whitepoint of the output
and patches the Lut to make sure it does.

When using the OptimiseMatrixShaper path through the code (which
can happen for 8 bit, non-alpha code), no such check is done. This
can lead to a mismatch in rendering, with a noticeable colour
shift at the whitepoint.

It's not possible to patch the tables used by OptimiseMatrixShaper
in the same way, so the best we can do is to spot such a mismatch
and to avoid using this path.